chore: mmigrated to .tsx, created types and interfaces as required
